refactor:#83 커스텀 훅, fetch 함수 분리
